Extract Nightscout secret generation into a shared helper function
Refactor the Nightscout secret generation logic by extracting it into a reusable helper function in `resources.js` and updating `tenant-composite-sync.js` to utilize this new function.
